Separate the reducer from store
